requirements.txt
streamlit
numpy
pandas
matplotlib
powerbiclient
openai
Pillow
requests